Continue building route and test

+++ routes/deleteUnit.js
@@ -7,11 +7,18 @@ router.post('/deleteUnit/:unitId', (req, res, next) => {
   const unitId = req.params.unitId;
 
   Unit.findById(unitId, (error, unit) => {
+    if (error) {
+      return next(error);
+    }
+
+    if (unit.movesRemaining < 1) {
+      return res.json(null);
+    }
+
     Unit.findByIdAndRemove(unitId, (error, res2) => {
       if (error) {
         return next(error);
       }
-
       return res.json(res2);
     });
   });
